html/semantics/interactive-elements/the-dialog-element/dialog-close.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>dialog element: close()</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#the-dialog-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<dialog id="d1">
  <p>foobar</p>
  <button>OK</button>
</dialog>
<dialog id="d2" open>
  <p>foobar</p>
  <button>OK</button>
</dialog>
<dialog id="d3" open>
  <p>foobar</p>
  <button>OK</button>
</dialog>
<script>
  var d1 = document.getElementById('d1'),
      d2 = document.getElementById('d2'),
      d3 = document.getElementById('d3'),
      t = async_test("close() fires a close event"),
      was_queued = false;

  d2.onclose = t.step_func_done(function(e) {
    assert_true(was_queued, "close event should be queued");
    assert_true(e.isTrusted, "close event is trusted");
    assert_false(e.bubbles, "close event doesn't bubble");
    assert_false(e.cancelable, "close event is not cancelable");
  });

  test(function(){
    assert_throws("INVALID_STATE_ERR", function() {
      d1.close();
    });
  }, "close() on a <dialog> that doesn't have an open attribute throws an InvalidStateError exception");

  test(function(){
    assert_true(d2.open);
    assert_equals(d2.returnValue, "");
    d2.close("closedialog");
    was_queued = true;
    assert_false(d2.hasAttribute("open"));
    assert_equals(d2.returnValue, "closedialog");
  }, "close() removes the open attribute and set the returnValue to the first argument");

  test(function(){
    assert_true(d3.open);
    assert_equals(d3.returnValue, "");
    d3.close();
    assert_false(d3.hasAttribute("open"));
    assert_equals(d3.returnValue, "");
  }, "close() without argument removes the open attribute and there's no returnValue");
</script>
